prep zod validation schemas for auth form changes
