+++ src/lib/types/auth/auth.types.ts
@@ -1,7 +1,7 @@
 import type { ChildrenProps } from '@/lib/types/global.types';
 import type { Dispatch, SetStateAction } from 'react';
 import type { IconType } from 'react-icons/lib';
-import type { Email, Login, OTP, Signup } from '../validation.types';
+import type { Email, Login, Signup } from '../validation.types';
 
 export type AuthLoginProviders = Readonly<'google' | 'github'>;
 export type AuthPageComponentProps = Readonly<{
@@ -23,17 +23,6 @@ export type TAuthenticationContext = {
 export type TAuthenticationContextProviderProps = {
 	children: React.ReactNode;
 };
-
-export type AuthContextReducerState = {
-	name: string | null;
-	email: Email['email'] | null;
-	otp: OTP['otp'] | null;
-};
-
-export type AuthContextReducerAction =
-	| { type: 'email'; email: string }
-	| { type: 'name'; name: string }
-	| { type: 'otp'; otp: string };
 
 export type AuthFormStates = 'register' | 'login' | 'otp';
